Second part of warning clears

+++ idk/projects/idk_opengl/system/FrameBufferManager.cpp
@@ -44,7 +44,7 @@ namespace idk::ogl
 			glFramebufferTexture(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0 + i, (GLuint)meta.textures[i]->ID(), 0);
 			buffers.push_back(GL_COLOR_ATTACHMENT0 + i);
 		}
-		glFramebufferRenderbuffer(GL_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_RENDERBUFFER, (GLuint)target->DepthBuffer());
+		glFramebufferRenderbuffer(GL_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_RENDERBUFFER, target->DepthBuffer());
 		glDrawBuffers(s_cast<GLsizei>(buffers.size()), buffers.data());
 
 		{
